rays working fine

+++ src/draw/cub3d_draw.h
@@ -6,7 +6,7 @@
 /*   By: lrocigno <lrocigno@student.42sp.org>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/14 21:35:28 by lrocigno          #+#    #+#             */
-/*   Updated: 2021/07/19 01:43:39 by lrocigno         ###   ########.fr       */
+/*   Updated: 2021/07/21 00:46:51 by lrocigno         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,8 +23,8 @@ void	pixel_put(t_img *img, int posx, int posy, int color);
 void	render_projection(t_img *f, t_wndw *wndw, t_world *wrld, t_actor *p);
 void	update_frame(t_wndw *wndw, t_world *wrld, t_actor *p);
 void	ray_cast(t_world *wrld, t_actor *player);
-t_ray	h_cast_ray(t_actor *player, t_world *wrld, double ang);
-t_ray	v_cast_ray(t_actor *player, t_world *wrld, double ang);
+t_ray	h_cast_ray(t_actor *player, t_world *wrld, float ang);
+t_ray	v_cast_ray(t_actor *player, t_world *wrld, float ang);
 void print_map(t_world *wrld, t_actor *player);
 
 #endif
